Se finaliza vista de cuenta del usuario
