+++ src/pages/accountpage/AccountPage.js
@@ -2,61 +2,26 @@ import "./accountpage.css";
 
 import React, { useEffect, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
+import { getUserFromToken, isAuthenticated } from "../../util/Auth";
 
 import Footer from "../../components/layouts/footer/Footer";
 import Navbar from "../../components/layouts/navbar/Navbar";
 
 const AccountPage = () => {
-  const linkStyle = { textDecoration: "none", color: "purple" };
+  const linkStyle = { textDecoration: "none", color: "gray" };
   const navigate = useNavigate();
   const [user, setUser] = useState(null);
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [message, setMessage] = useState(null);
 
   // Función para obtener los datos del usuario actual
-  const fetchUserData = async () => {};
-
   useEffect(() => {
-    fetchUserData();
+    const userFromToken = getUserFromToken();
+
+    if (userFromToken) {
+      setUser(userFromToken);
+    } else {
+      navigate("/signin");
+    }
   }, []);
-
-  // Función para manejar el cierre de sesión
-  const handleLogout = async () => {
-    // try {
-    //   await auth.signOut();
-    //   navigate("/signin");
-    // } catch (error) {
-    //   console.error("Error al cerrar sesión:", error.message);
-    // }
-  };
-
-  // Función para manejar el cambio de contraseña
-  const handleChangePassword = async (e) => {
-    e.preventDefault();
-
-    // if (password !== confirmPassword) {
-    //   setMessage("Las contraseñas no coinciden");
-    //   return;
-    // }
-
-    // const user = auth.currentUser;
-    // const credential = firebase.auth.EmailAuthProvider.credential(
-    //   user.email,
-    //   user.password
-    // );
-
-    // try {
-    //   await user.reauthenticateWithCredential(credential);
-    //   await user.updatePassword(password);
-    //   setMessage("Contraseña cambiada correctamente");
-    //   setPassword("");
-    //   setConfirmPassword("");
-    // } catch (error) {
-    //   setMessage(error.message);
-    //   console.error("Error al cambiar la contraseña:", error.message);
-    // }
-  };
 
   return (
     <section>
@@ -65,68 +30,33 @@ const AccountPage = () => {
         {user ? (
           <div className="my-5 py-5">
             <div className="row container mx-auto">
-              <div className="text-center mt-3 pt-5 col-lg-6 col-md-12 col-sm-12">
-                <h3 className="font-weight-bold">Información de Cuenta</h3>
+              <div className="account-container text-center mt-3 pt-5">
+                <img
+                  src="/usericon.png"
+                  alt="User Profile"
+                  className="user-image"
+                />
+
+                <h3 className="title">{user.name}</h3>
+                <p className="subtitle">
+                  <strong>Rol:</strong> {user.role}
+                </p>
+
                 <hr className="mx-auto" />
                 <div className="account-info">
-                  <p style={linkStyle}>{user.email}</p>
-                  <p>
-                    <Link to="/cart" style={linkStyle} id="order-btn">
-                      Tus pedidos
-                    </Link>
+                  <p style={linkStyle}>
+                    <strong>Email:</strong> {user.email}
                   </p>
-                  <p>
-                    <a
-                      href="/signin"
-                      id="logout-btn"
-                      onClick={handleLogout}
-                      style={linkStyle}
-                    >
-                      Cerrar Sesión
-                    </a>
+                  <p style={linkStyle}>
+                    <strong>Telefono:</strong> 555-555-5555
                   </p>
+                  <button
+                    className="edit-btn"
+                    onClick={() => navigate("/edit-account")}
+                  >
+                    Editar cuenta
+                  </button>
                 </div>
-              </div>
-              <div className="col-lg-6 col-md-12 col-sm-12">
-                <form id="account-form" onSubmit={handleChangePassword}>
-                  <h3>Cambiar Contraseña</h3>
-                  <hr className="mx-auto" />
-                  <div className="form-group">
-                    <label>Contraseña</label>
-                    <input
-                      type="password"
-                      className="form-control"
-                      id="account-password"
-                      name="password"
-                      placeholder="Contraseña"
-                      value={password}
-                      onChange={(e) => setPassword(e.target.value)}
-                      required
-                    />
-                  </div>
-                  <div className="form-group">
-                    <label>Confirmar Contraseña</label>
-                    <input
-                      type="password"
-                      className="form-control"
-                      id="account-password-confirm"
-                      name="confirmPassword"
-                      placeholder="Confirmar Contraseña"
-                      value={confirmPassword}
-                      onChange={(e) => setConfirmPassword(e.target.value)}
-                      required
-                    />
-                  </div>
-                  <div className="form-group">
-                    <input
-                      type="submit"
-                      className="btn"
-                      value="Confirmar"
-                      id="change-pass-btn"
-                    />
-                  </div>
-                </form>
-                {message && <p className="text-danger">{message}</p>}
               </div>
             </div>
           </div>
