+++ src/pages/editaccountpage/editaccountpage.css
@@ -1,0 +1,20 @@
+#account-form{
+    width: 50%;
+    margin: 25px auto;
+    text-align: center;
+    padding: 20px;
+}
+
+#account-form input{
+    margin: 5px auto;
+}
+
+#account-form #change-pass-btn{
+    color: #fff;
+    font-weight: 900;
+    background-color:var(--color-primario2);
+}
+.account-info #order-btn,#logout-btn{
+    color: var(--color-primario2);
+    text-decoration: none;
+}
